// tlk-frontend19/src/app/api/model/dokumentenstapel.ts
/**
 * TheLionKing
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface Dokumentenstapel { 
    id?: string;
    vorgangId?: string;
    stapelName?: string;
    uploadFilename?: string;
    uploadPdf?: string;
    originalFilename?: string;
    pdfStorageRef?: string;
    seitenAnzahl?: number;
    status?: string;
    createdAt?: string;
}